@@ -3,25 +3,29 @@ package com.loysc.zzangco.kirikiri_snu.activity;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.design.widget.NavigationView;
-import android.support.v4.view.GravityCompat;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBarDrawerToggle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+
+import androidx.appcompat.app.ActionBarDrawerToggle;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+import androidx.core.view.GravityCompat;
+import androidx.drawerlayout.widget.DrawerLayout;
+import androidx.recyclerview.widget.RecyclerView;
+
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.webkit.WebView;
+import android.view.animation.Animation;
+import android.view.animation.AnimationUtils;
+import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+import android.widget.ViewFlipper;
 
+import com.google.android.material.navigation.NavigationView;
 import com.loysc.zzangco.kirikiri_snu.R;
 
 import java.util.ArrayList;
 
-import com.loysc.zzangco.kirikiri_snu.common.MemberAdapter;
 import com.loysc.zzangco.kirikiri_snu.common.MemberViewItem;
 
 public class OlcActivity extends AppCompatActivity
@@ -31,23 +35,29 @@ public class OlcActivity extends AppCompatActivity
     private RecyclerView.LayoutManager layoutManager;
     private ArrayList<MemberViewItem> items = new ArrayList<MemberViewItem>();
 
-    private MemberAdapter adapter;
+    //private MemberAdapter adapter;
     private String granNumber = "00";
     private LinearLayout llOlcIntro;
     private LinearLayout llOlcChart;
     private LinearLayout llOlcHistory;
     private LinearLayout llOlcLaw;
-    private LinearLayout llOlcOffice;
-    private WebView webViewMap;
+    private LinearLayout llOlcStaff;
+
 
     private TextView olcOfficePhone,olcOfficeEmail;
+
+    private ViewFlipper vfSlider;
+    private ImageView imgBanner1,imgBanner2,imgBanner3,imgBanner5,imgBanner7,imgBanner8,imgBanner9;
+
+    private Animation slide_out_left,slide_in_right;
+    private Animation slide_in_left,slide_out_right;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_olc_list);
         toolbar = (Toolbar) findViewById(R.id.toolbar);
-        adapter = new MemberAdapter(this,items);
+        //adapter = new MemberAdapter(this,items);
         setSupportActionBar(toolbar);
 
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
@@ -63,10 +73,8 @@ public class OlcActivity extends AppCompatActivity
         llOlcChart = (LinearLayout)findViewById(R.id.llOlcChart);
         llOlcHistory = (LinearLayout)findViewById(R.id.llOlcHistory);
         llOlcLaw = (LinearLayout)findViewById(R.id.llOlcLaw);
-        llOlcOffice = (LinearLayout)findViewById(R.id.llOlcOffice);
-        webViewMap = (WebView)findViewById(R.id.webViewMap);
-        webViewMap.loadUrl("file:///android_asset/olcMap.html");
-        webViewMap.getSettings().setJavaScriptEnabled(true);
+        llOlcStaff = (LinearLayout)findViewById(R.id.llOlcStaff);
+
 
         olcOfficePhone = (TextView)findViewById(R.id.olcOfficePhone);
 
@@ -86,14 +94,101 @@ public class OlcActivity extends AppCompatActivity
             }
         });
 
-        getSupportActionBar().setTitle("OLC란?");
+        getSupportActionBar().setTitle("회장단소개");
         MainActivity.instance.asyncDialog.dismiss();
+
+        vfSlider = (ViewFlipper)findViewById(R.id.vfSlider);
+        imgBanner1 = (ImageView)findViewById(R.id.imgBanner1);
+        imgBanner2 = (ImageView)findViewById(R.id.imgBanner2);
+        imgBanner3 = (ImageView)findViewById(R.id.imgBanner3);
+        imgBanner5 = (ImageView)findViewById(R.id.imgBanner5);
+        imgBanner7 = (ImageView)findViewById(R.id.imgBanner7);
+        imgBanner8 = (ImageView)findViewById(R.id.imgBanner8);
+        imgBanner9 = (ImageView)findViewById(R.id.imgBanner9);
+
+        imgBanner1.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.gjec_url)));
+                startActivity(intent);
+            }
+        });
+        imgBanner2.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.anysho_url)));
+                startActivity(intent);
+            }
+        });
+        imgBanner3.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.inettv_url)));
+                startActivity(intent);
+            }
+        });
+
+        imgBanner5.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.ysk_url)));
+                startActivity(intent);
+            }
+        });
+        imgBanner7.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.vhc_url)));
+                startActivity(intent);
+            }
+        });
+        imgBanner8.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.modoo_url)));
+                startActivity(intent);
+            }
+        });
+        imgBanner9.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(Intent.ACTION_VIEW, Uri.parse(getString(R.string.softmill_url)));
+                startActivity(intent);
+            }
+        });
+
+        slide_in_left = AnimationUtils.loadAnimation(this,android.R.anim.slide_in_left );
+        slide_out_right = AnimationUtils.loadAnimation(this, android.R.anim.slide_out_right );
+
+        slide_out_left = AnimationUtils.loadAnimation(this,R.anim.ani_translate_l );
+        slide_in_right = AnimationUtils.loadAnimation(this,R.anim.ani_translate_r );
+
+        vfSlider.setInAnimation(slide_in_right);
+        vfSlider.setOutAnimation(slide_out_left);
+    }
+
+    @Override
+    protected void onResume() {
+        super.onResume();
+        //int firstViewInt = ZZangcoUtility.randomRange(0,2);
+
+        vfSlider.setDisplayedChild(0);
+        vfSlider.startFlipping();
+
+        vfSlider.setFlipInterval(2500);
     }
 
     private void callPhone(){
         //phoneNumber = "02-828-3962";
-        String olc_office_phone = getString(R.string.olc_office_phone);
+        String olc_office_phone = getString(R.string.olp_office_phone);
         Intent intent = new Intent(Intent.ACTION_DIAL, Uri.parse("tel:"+olc_office_phone));
+        startActivity(intent);
+
+    }
+    private void callPhone2(){
+        //phoneNumber = "02-705-8017";
+        String olp_office_phone = getString(R.string.olp_office_phone2);
+        Intent intent = new Intent(Intent.ACTION_DIAL, Uri.parse("tel:"+olp_office_phone));
         startActivity(intent);
 
     }
@@ -155,19 +250,19 @@ public class OlcActivity extends AppCompatActivity
             llOlcChart.setVisibility(View.GONE);
             llOlcHistory.setVisibility(View.GONE);
             llOlcLaw.setVisibility(View.GONE);
-            llOlcOffice.setVisibility(View.GONE);
+            llOlcStaff.setVisibility(View.GONE);
 
-            toolbar.setTitle("OLC란?");
+            toolbar.setTitle("회장단소개");
         } else if (id == R.id.olcChart) {
 
             llOlcIntro.setVisibility(View.GONE);
             llOlcChart.setVisibility(View.VISIBLE);
             llOlcHistory.setVisibility(View.GONE);
             llOlcLaw.setVisibility(View.GONE);
-            llOlcOffice.setVisibility(View.GONE);
+            llOlcStaff.setVisibility(View.GONE);
 
             toolbar.setTitle("조직도");
-        } else if (id == R.id.olcHistory) {
+        /*} else if (id == R.id.olcHistory) {
 
             llOlcIntro.setVisibility(View.GONE);
             llOlcChart.setVisibility(View.GONE);
@@ -175,26 +270,26 @@ public class OlcActivity extends AppCompatActivity
             llOlcLaw.setVisibility(View.GONE);
             llOlcOffice.setVisibility(View.GONE);
 
-            toolbar.setTitle("OLC 연혁");
+            toolbar.setTitle("OLC 연혁");*/
         } else if (id == R.id.olcLaw) {
 
             llOlcIntro.setVisibility(View.GONE);
             llOlcChart.setVisibility(View.GONE);
             llOlcHistory.setVisibility(View.GONE);
             llOlcLaw.setVisibility(View.VISIBLE);
-            llOlcOffice.setVisibility(View.GONE);
+            llOlcStaff.setVisibility(View.GONE);
 
-            toolbar.setTitle("회칙");
-        } else if (id == R.id.olcOffice) {
+            toolbar.setTitle("회칙소개");
+        } /*else if (id == R.id.olcStaff) {
 
             llOlcIntro.setVisibility(View.GONE);
             llOlcChart.setVisibility(View.GONE);
             llOlcHistory.setVisibility(View.GONE);
             llOlcLaw.setVisibility(View.GONE);
-            llOlcOffice.setVisibility(View.VISIBLE);
+            llOlcStaff.setVisibility(View.VISIBLE);
 
-            toolbar.setTitle("회칙");
-        }
+            toolbar.setTitle("임원진 소개");
+        }*/
 
         DrawerLayout drawer = (DrawerLayout) findViewById(R.id.drawer_layout);
         drawer.closeDrawer(GravityCompat.START);
