@@ -1,13 +1,21 @@
 package com.loysc.zzangco.kirikiri_snu.activity;
 
+import android.Manifest;
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
 import android.os.Bundle;
 import android.os.Handler;
 import android.os.Message;
-import android.support.v4.util.Pair;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.core.util.Pair;
+
 import android.telephony.TelephonyManager;
 import android.util.Log;
 import android.view.View;
@@ -25,10 +33,12 @@ import java.io.BufferedOutputStream;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.InputStream;
+import java.lang.ref.WeakReference;
 import java.util.Arrays;
 import java.util.List;
 
 import com.loysc.zzangco.kirikiri_snu.common.BoardInfo;
+import com.loysc.zzangco.kirikiri_snu.common.CheckInfo;
 import com.loysc.zzangco.kirikiri_snu.common.DataBase;
 import com.loysc.zzangco.kirikiri_snu.common.MemberInfo;
 import com.loysc.zzangco.kirikiri_snu.common.ScheduleInfo;
@@ -42,67 +52,135 @@ import com.loysc.zzangco.kirikiri_snu.vo.MemberVo;
 import com.loysc.zzangco.kirikiri_snu.vo.ResultVo;
 import com.loysc.zzangco.kirikiri_snu.vo.ScheduleVo;
 
+import static com.loysc.zzangco.kirikiri_snu.activity.MainActivity.USER_PHONE_NUM;
+
 public class SplashActivity extends AppCompatActivity implements HttpConnectionThread {
     private ImageView imgSplash;
 
     private Animation animationTL;
 
-    private Pair<View,String> pair1;
+    private Pair<View, String> pair1;
 
     private MemberInfo memberInfo;
 
     private BoardInfo boardInfo;
 
+    private CheckInfo checkinfo;
+
     private ScheduleInfo scheduleInfo;
 
     private HttpConnection httpConnection;
-    private HttpConnectionNoHandler httpConnectionNoHandler;
-    Handler handler;
+    private NonLeakHandler handler = new NonLeakHandler(this);
+
     String myNumber = null;
+
+    String userConfirm = null;
+
+    private static int MY_PERMISSIONS_CALL_PHONE = 0;
+    public static int USER_CONFIRM = 7;
+    public static String IS_OK = "isOk";
+
+    private static final class NonLeakHandler extends Handler {
+        private final WeakReference<SplashActivity> ref;
+
+        NonLeakHandler(SplashActivity act) {
+            ref = new WeakReference<>(act);
+        }
+
+        @Override
+        public void handleMessage(Message msg) {
+            SplashActivity act = ref.get();
+            if (act != null) {
+                Intent intent = new Intent(act.getApplicationContext(), MainActivity.class);
+                intent.putExtra(USER_PHONE_NUM, act.myNumber);
+                //startActivity(intent,option.toBundle());
+                act.startActivity(intent);
+                act.finish();
+            }
+        }
+    }
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_splash);
-
-        imgSplash = (ImageView)findViewById(R.id.imgSplash);
-
         startWork();
+        imgSplash = (ImageView) findViewById(R.id.imgSplash);
 
-        animationTL = AnimationUtils.loadAnimation(this,R.anim.ani_translate_r);
+        animationTL = AnimationUtils.loadAnimation(this, R.anim.ani_translate_r);
 
         imgSplash.setVisibility(View.VISIBLE);
         imgSplash.startAnimation(animationTL);
 
-        pair1 = Pair.create(findViewById(R.id.imgSplash),"mainImg");
+        //Glide.with(this).load(R.drawable.kirikiri).into(imgSplash);
 
-        handler = new Handler(){
-            @Override
-            public void handleMessage(Message msg) {
-                //ActivityOptionsCompat option = ActivityOptionsCompat.makeSceneTransitionAnimation(SplashActivity.this,pair1);
-                Intent intent = new Intent(getApplicationContext(),MainActivity.class);
-                //startActivity(intent,option.toBundle());
-                startActivity(intent);
-                finish();
+
+        pair1 = Pair.create(findViewById(R.id.imgSplash), "mainImg");
+
+        getCheckDatastore();
+        checkinfo.openc();
+        userConfirm = checkinfo.getCheck();
+        checkinfo.close();
+
+
+        if (userConfirm.equals("Y")) {
+            goProgress();
+        }else{
+            Intent intent = new Intent(getApplicationContext(), InAppInfoActivity.class);
+            startActivityForResult(intent, USER_CONFIRM);
+        }
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+        if(resultCode == RESULT_OK) {
+            if (requestCode == USER_CONFIRM) {
+                boolean returnValue = data.getBooleanExtra(IS_OK, false);
+                if (returnValue) {
+                    getCheckDatastore();
+                    checkinfo.openc();
+                    checkinfo.setCheck("Y");
+                    checkinfo.close();
+
+                    permissonCheckFu();
+                } else {
+                    getCheckDatastore();
+                    checkinfo.openc();
+                    checkinfo.setCheck("N");
+                    checkinfo.close();
+
+                    finish();
+                }
+
             }
-        };
+        }
+    }
+    @Override
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        super.onRequestPermissionsResult(requestCode, permissions, grantResults);
+        if(requestCode == MY_PERMISSIONS_CALL_PHONE){
+            goProgress();
+        }
+    }
+
+    @SuppressLint("HardwareIds")
+    private void goProgress(){
+
+        if (ContextCompat.checkSelfPermission(this, Manifest.permission.READ_PHONE_STATE) == PackageManager.PERMISSION_GRANTED) {
+            TelephonyManager mgr = (TelephonyManager) getSystemService(Context.TELEPHONY_SERVICE);
+            if (mgr != null) {
+                myNumber = mgr.getLine1Number();
+                myNumber = myNumber.replace("+82", "0");
+                myNumber = myNumber.replace("-", "");
+                myNumber = myNumber.trim();
+            }
 
 
-        TelephonyManager mgr = (TelephonyManager) getSystemService(Context.TELEPHONY_SERVICE);
-
-
-
-        try{
-            myNumber = mgr.getLine1Number();
-
-            Log.e("zzangco","phoneNumber ["+myNumber+"]");
-           // myNumber = myNumber.replace("+", "");
-            myNumber = myNumber.replace("+82", "0");
-            myNumber = myNumber.replace("-", "");
-            myNumber = myNumber.trim();
-        }catch(Exception e){}
+        }
 
         if(checkPhoneNumber(myNumber)){
-            //handler.sendEmptyMessageDelayed(0,1000);
+            //handler.sendEmptyMessageDelayed(0,3000);
             String url = getString(R.string.mainUrl);
             url += getString(R.string.getInfo);
             url += "?" + ZZangcoUtility.CHECKUSER_PARAM + "=" + myNumber;
@@ -110,7 +188,9 @@ public class SplashActivity extends AppCompatActivity implements HttpConnectionT
             Log.e("zzangco","testURL=["+ url + "]");
             httpConnection = new HttpConnection(this, url,HttpConnection.URLTYPE.SEND);
             httpConnection.start();
+            Log.e("zzangco","전화번호 있음");
         }else{
+            // handler.sendEmptyMessageDelayed(0,3000);
             Log.e("zzangco","phoneNumber ["+myNumber+"]");
             Log.e("zzangco","로그인 실패함");
 
@@ -122,15 +202,20 @@ public class SplashActivity extends AppCompatActivity implements HttpConnectionT
             httpConnection = new HttpConnection(this, url,HttpConnection.URLTYPE.SEND);
             httpConnection.start();
 
+        }
+    }
+    private void permissonCheckFu(){
+        if (ContextCompat.checkSelfPermission(this,Manifest.permission.READ_PHONE_STATE) != PackageManager.PERMISSION_GRANTED) {
 
+            if (ActivityCompat.shouldShowRequestPermissionRationale(this,Manifest.permission.READ_PHONE_STATE)) {
 
+            }else {
+                ActivityCompat.requestPermissions(this,new String[]{Manifest.permission.READ_PHONE_STATE},MY_PERMISSIONS_CALL_PHONE);
+            }
         }
 
 
-
-
     }
-
 
     private  void startWork(){
         if(!isCheckDB()){
@@ -142,11 +227,13 @@ public class SplashActivity extends AppCompatActivity implements HttpConnectionT
         getDatastore();
         memberInfo.openc();
 
+
         if(memberInfo.checkPhoneNumber(number)){
             return true;
         }else {
             return false;
         }
+
 
         //return true;
     }
@@ -175,6 +262,12 @@ public class SplashActivity extends AppCompatActivity implements HttpConnectionT
         }
 
         return boardInfo;
+    }
+
+    private void getCheckDatastore(){
+        if(null == checkinfo){
+            checkinfo = new CheckInfo(getApplication());
+        }
     }
 
     private boolean addBoard(BoardVo boardVo){
@@ -324,8 +417,8 @@ public class SplashActivity extends AppCompatActivity implements HttpConnectionT
                 handler = null;
                 Intent intent = new Intent(getApplicationContext(),PopupActivity.class);
                 //startActivity(intent,option.toBundle());
-                intent.putExtra(PopupActivity.TITLE,"OLC 멤버 체크");
-                intent.putExtra(PopupActivity.CONTENT,"OLC 멤버로 확인 되지 않습니다. \n OLC 사무국(02-3289-4214)으로 확인 부탁드립니다.");
+                intent.putExtra(PopupActivity.TITLE,"SPARC 멤버 체크");
+                intent.putExtra(PopupActivity.CONTENT,"SPARC 멤버로 확인 되지 않습니다. \n SPARC 행정실(02-880-6251)으로 확인 부탁드립니다.");
                 startActivity(intent);
                 finish();
             }else{
@@ -343,8 +436,8 @@ public class SplashActivity extends AppCompatActivity implements HttpConnectionT
             Log.e("zzangco", "error:" + e.getMessage());
             e.printStackTrace();
         }finally {
-            if(null != handler){
-                handler.sendEmptyMessageDelayed(0,3000);
+            if(null != handler) {
+                handler.sendEmptyMessageDelayed(0, 3000);
             }
         }
     }
