@@ -1,10 +1,11 @@
 package com.loysc.zzangco.kirikiri_snu.common;
 
-import android.support.v7.widget.CardView;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.cardview.widget.CardView;
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.loysc.zzangco.kirikiri_snu.R;
 
